ZKPs moved to crypto/zkp. Proofs moved into crypto/zkp/primitives and crypto/zkp/schemes. Functions to demonstrate how ZKPs should be executed added or equipped with comments.
