+++ crypto/zkp/schemes/pseudonymsys/org_nym_gen.go
@@ -22,7 +22,7 @@ import (
 	"fmt"
 	"github.com/xlab-si/emmy/crypto/common"
 	"github.com/xlab-si/emmy/crypto/dlog"
-	"github.com/xlab-si/emmy/crypto/dlogproofs"
+	"github.com/xlab-si/emmy/crypto/zkp/primitives/dlogproofs"
 	"math/big"
 )
