+++ test/qr_test.go
@@ -25,9 +25,34 @@ import (
 	"github.com/xlab-si/emmy/log"
 	"math/big"
 	"testing"
+	"github.com/xlab-si/emmy/crypto/zkp/primitives/qrproofs"
 )
 
 func TestQRProof(t *testing.T) {
+	dlog := config.LoadDLog("pseudonymsys")
+
+	secret := common.GetRandomInt(dlog.OrderOfSubgroup)
+	groupOrder := new(big.Int).Sub(dlog.P, big.NewInt(1))
+	g1, _ := common.GetGeneratorOfZnSubgroup(dlog.P, groupOrder, dlog.OrderOfSubgroup)
+	y1, _ := dlog.Exponentiate(g1, secret)
+
+	proved := qrproofs.ProveQR(y1, dlog)
+
+	assert.Equal(t, proved, true, "QR proof does not work correctly")
+}
+
+func TestQNRProof(t *testing.T) {
+	qr := config.LoadQR("qrsmall")
+	y, _ := new(big.Int).SetString("12104178836609367680573806127379667907452906068454411069540554013287299560619180831355030260792398116234585889489580787593876656468666620239704236202828625502145041222925221901747074982936094534863675673392263672910937547483238701395223837362913804135100562910313510524388491518118503153440479519531614792845582743827952578371105856240886035300256188119597496494955532842085616018276731618147827132295654573847443973865791627336995666490060797108039052526091724109438973877494599020695354285996982775389058505616271009101591441286538176405813622092518363027538767195421845003207037113604997900115835538908295151715618", 10)
+	proved, err := qrproofs.ProveQNR(y, qr)
+	if err != nil {
+		t.Errorf("Error when proving QNR")
+	}
+
+	assert.Equal(t, proved, true, "QNR proof does not work correctly")
+}
+
+func TestQRProofGRPC(t *testing.T) {
 	log.TurnOff() // otherwise too many communication messages
 	dlog := config.LoadDLog("pseudonymsys")
 	y1 := common.GetRandomInt(dlog.P)
@@ -42,10 +67,10 @@ func TestQRProof(t *testing.T) {
 		t.Errorf("Error when proving y is QR")
 	}
 
-	assert.Equal(t, proved, true, "QR proof does not work correctly")
+	assert.Equal(t, proved, true, "QR proof over gRPC does not work correctly")
 }
 
-func TestQNRProof(t *testing.T) {
+func TestQNRProofGRPC(t *testing.T) {
 	log.TurnOff() // otherwise too many communication messages
 	qr := config.LoadQR("qrsmall")
 	y, _ := new(big.Int).SetString("12104178836609367680573806127379667907452906068454411069540554013287299560619180831355030260792398116234585889489580787593876656468666620239704236202828625502145041222925221901747074982936094534863675673392263672910937547483238701395223837362913804135100562910313510524388491518118503153440479519531614792845582743827952578371105856240886035300256188119597496494955532842085616018276731618147827132295654573847443973865791627336995666490060797108039052526091724109438973877494599020695354285996982775389058505616271009101591441286538176405813622092518363027538767195421845003207037113604997900115835538908295151715618", 10)
@@ -60,5 +85,5 @@ func TestQNRProof(t *testing.T) {
 		t.Errorf("Error when proving y is QNR")
 	}
 
-	assert.Equal(t, proved, true, "QNR proof does not work correctly")
+	assert.Equal(t, proved, true, "QNR proof over gRPC does not work correctly")
 }
